crypto: caam - add functionality used by the caam_dma driver
commit  12695aeb32ee939917a07a0c468c1624d452ce12 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The caam_dma is a memcpy DMA driver based on the DMA functionality of
the CAAM hardware block. It creates a DMA channel for each JR of the
CAAM. This patch adds functionality that is used by the caam_dma that is
not yet part of the JR driver.

Signed-off-by: Radu Alexe <radu.alexe@nxp.com>
Signed-off-by: Horia Geantă <horia.geanta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>