x86/resctrl: Allow overflow/limbo handlers to be scheduled on any-but cpu
commit 57da9317e9f978be67c75ec86f6d920505a11931 from
git@git.assembla.com:cavium/WindRiver.linux.git

When a cpu is taken offline resctrl may need to move the overflow or
limbo handlers to run on a different cpu. When splitting the
cpuhotplug callbacks up between arch and fs we now make this decision
before the die-ing cpu has been removed from the domain cpumasks.

resctrl_offline_cpu() is told which cpu is going away, so it
can ask the handler setup functions to exclude the specified cpu.
Use -1 as an any-cpu-will-do value for the normal case.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>