MLK-22158 drm/imx: lcdif: force 'mode_changed' for fb pix fmt change
commit  726892c6b387aea04fb44875e67042a6c59a970d from
https://github.com/nxp-imx/linux-imx lf-6.12.y

According to the LCDIF function definition, the pixel format
configuration should not be done during LCDIF is running. So
if finding pixel format change is requested, the 'mode_changed'
needs to be set to make sure the commit become to be a full
modeset.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>