drm/imx: dpu: crtc: Do not send out bogus vblank event in ->atomic_disable()
commit  f04c207d66bf99484716e4bba31ab92a46db1728 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

When a full modeset is needed, the CRTC could be totally disabled or
enabled/re-enabled after the modeset.  If it's re-enabled, a vblank
event would be sent during the CRTC enablement procedure.  So, a bogus
event should be killed in the ->atomic_disable() callback.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>