LF-794 arm64: dts: imx8qm: Add firmware-name properties for HDMI and DP
commit  27e7d80dd0c8b7b38acca480473b45b3c89cdcae from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Add the display port firmware name property and set it to dpfw.bin for
the lpddr4 validation board, and the hdmi firmware name property and
set it to hdmitxfw.bin for the mek board. This tells the driver to load
right the firmware based on the compatible property.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Sandor Yu <sandor.yu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>