soc: fsl: dpio: fix qbman alignment error in the virtualization context
commit  22e92183b50d9eb8c0d8ce3fdca4be8a77dea789 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

When running as a guest, under KVM, the CENA region is mapped
as device memory, so uncacheable. When the memory is mapped
as device memory, the unaligned accesses are not allowed.
Memcpy is optimized to transfer 8 bytes at a time regardless
of the start address and might cause alignment issues.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>