LF-3029 arm64: dts: imx8dxl: add v2x subsystem
commit  74a6905b83a3a992aa710e02e92c83467153eeff from
https://github.com/nxp-imx/linux-imx lf-6.12.y

This patch adds the V2X subsystem in which the different
MUs to access the V2X are listed. Without this patch the DXL
fails to enter KS1 as V2X resource is left ON and suspend
power is very high.

Signed-off-by: Stéphane Dion <stephane.dion_1@nxp.com>
Signed-off-by: Nitin Garg <nitin.garg@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>