LF-1411-2 drm/bridge: adv7511: implement drm_bridge detach
commit  8f86df7cbcdfd4d4908af8070649d44574d644fc from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The detach callback was not implemented in adv7511 driver, so implement
it now, in order to have a correct cleanup mechanism. Also, move the
dsi detach into drm_bridge detach, so the operations are done in the
correct reverse order.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
[ Liu Ying: Fix build break by using adv->info->has_dsi to upgrade to next-20231129 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>