x86/resctrl: Move closid/rmid matching and setting to use helpers
commit d42aeba1631708e9984eb728d781be8d47a4d882 from
git@git.assembla.com:cavium/WindRiver.linux.git

When switching tasks, the closid and rmid that the new task should
use are store in struct task_struct. For x86 the closid known by resctrl,
the value in task_struct, and the value written to the CPU register are
all the same thing.

MPAM's CPU interface has two different partid's one for data accesses
the other for instruction fetch. Storing resctrl's closid value in
struct task_struct implies the arch code knows whether resctrl is using
CDP, and there are no other users.

Move the matching and setting of the struct task_struct properties
to use helpers. This allows arm64 to use this to hold the hardware
format of the register, instead of having to convert it each time.
__rdtgroup_move_task()s use of READ_ONCE()/WRITE_ONCE() ensures torn
values aren't seen as another CPU may schedule the task being moved
while the value is being changed. MPAM has an additional corner-case
here as the PMG bits extend the PARTID space. If the scheduler sees a
new-closid but old-rmid, the task will dirty an rmid that the limbo code
is not watching causing an inaccurate count. RDT's rmid are independent
values, so the limbo code will still be watching the old-rmid in this
circumstance.
To avoid this, arm64 needs both the closid/rmid WRITE_ONCE()d together.
Both values must be provided together.
Because MPAM's rmid values are not unique, the closid must be provided
when matching the rmid.

CC: Valentin Schneider <valentin.schneider@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>