MLK-24805-1: drm/panel: rm67191: Add support for rm67199
commit  ee51f832ce44b8b44434094b8f3ff377212cb8f6 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

There is a new and improved version of this panel, which is called
rm67199, and is the same as rm67191 (in size, aspect and resolution).
Since, only the DSI initialization sequence differs, I added this
into the existing driver for rm67191.
So:
- compacted the DSI commands into 4 per line so they will not consume
too many lines of code
- updated the prepare sequence, since the instructions for this panel
  were provided (the sleeps for rm67191 were just guesses)
- added platform data to determine which version of the panel is used
  with this driver

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
[ Liu Ying: rebase upon next-20240618 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>