LF-3024: drm/imx/dcss: allow primary plane to be less than crtc size
commit  d8fe1b4b6b7c677363614371ef1fe1dcf24f4210 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Currently, the primary plane has to cover the entire CRTC area. But, there are
situations when we want the primary plane to be smaller, to save bandwidth.
Since the HW allows it, activate it in the driver as well.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Acked-by: Mirela Rabulea <mirela.rabulea@nxp.com>
[ Liu Ying: Fix conflicts to upgrade to next-20210514 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>