LF-3319 net: stmmac: enable rx_clk when killing the vlan
commit  eedd4a93749a6ead84771de13f0047f7208a2de9 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

when killing the vlan, the rx_clk from the phy need to be enabled.
The patch checks the phy's status before killing the vlan. If the phy
is stopped, enable it, remove the vlan, and then stop the phy.

Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>