dpaa2-eth: Set congestion group before the Rx queues become operational
commit  d3bb261bc4c60222302a82d1b37f24b79e947d75 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Setting  the CGID for the first time or dynamically changing it
with a different one  on a FQ that has been in service (has frames in it)
is not supported, the byte/frame counts in both the old and new CGRs may be
corrupted if this is done.

This patch adds the configuration of the CGR(congestion group record)
for the network interface RX queues, in the probe function before the queues
become operational,  to prevent the corruption of the aforementioned counter.

Signed-off-by: Radu Bulie <radu-andrei.bulie@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>