arm_mpam: Add a helper to touch an MSC from any CPU
commit 7d63dbdd99499e559c85795ee777e83456a1e4eb from
git@git.assembla.com:cavium/WindRiver.linux.git

Resetting RIS entries from the cpuhp callback is easy as the
callback occurs on the correct CPU. This won't be true for any other
caller that wants to reset or configure an MSC.

Add a helper that schedules the provided function if necessary.
Prevent the cpuhp callbacks from changing the MSC state by taking the
cpuhp lock.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>