ACPI / PPTT: Add a helper to fill a cpumask from a cache_id
commit 2310559b8520d61999d81b684faaa21086e99f21 from
git@git.assembla.com:cavium/WindRiver.linux.git

MPAM identifies CPUs by the cache_id in the PPTT cache structure.

The driver needs to know which CPUs are associated with the cache,
the CPUs may not all be online, so cacheinfo does not have the
information.

Add a helper to pull this information out of the PPTT.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>