arm64: dts: freescale: Switch from clock-indices to bit-offsets
commit  bd05b20cd3a178b34270948624950a2570da3482 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

This is a temporary fix after rebasing the linux-nxp/dts/imx8
on top of next-20210514. The reason we're doing the switch is to
keep the dts working. The upstream version of the dts files already
use clock-indices. The NXP's tree needs to become ready to do so
too.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>