MLK-25591-03: pxp: Kconfg: remove dependency of 32 bits ARM platform
commit  4a4602a19d0b3831f639867991bd0c8798d795b8 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Due to PXP is used in iMX6, iMX7 Serial platform and they are 32 bits arm
platform, so add the dependency for PXP driver.

In iMX8ULP, it also integrate PXP and it's 64 bits arm platform, so remove
the dependency for it.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by  : Robby.Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>