arm64: mpam: Add helpers to change a tasks and cpu mpam partid/pmg values
commit 72e04a265c6df5e10032727250922634b2c94304 from
git@git.assembla.com:cavium/WindRiver.linux.git

Care must be taken when modifying the partid and pmg of a task, as
writing these values may race with the task being scheduled in, and
reading the modified values.

Add helpers to set the task properties, and the cpu default value,
and add the plumbing to the mpam driver that lets resctrl use them.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>