LF-4502 drm/imx: dpu: dpu-crc: Avoid out-of-bounds access to array 'params' for ROI
commit  a02e70b43dfe87acf52cdb8c067e6eeccaf2d597 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The number of Region Of Interest(ROI) parameters(delimited by comma) is 4,
so we should stop getting parameters if the iterator i is greater than 3
where i starts from 0.  We cache the parameters in the array 'params' whose
size is also 4.  This patch fixes the check for the iterator to avoid
out-of-bounds access to the array 'params'.

This fixes Coverity issue: CID 12358615.

Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>