kobject: Add kset_get_next_obj() to allow a kset to be walked
commit 41d3dadb7d6a5ca6951aad81c3cbf6e08f70f1af from
git@git.assembla.com:cavium/WindRiver.linux.git

To expose iommu_groups via the resctrl filesystem, the resctrl driver
needs to be able to walk the list of iommu_groups. These are exposed
via sysfs as a kset.

Add kset_get_next_obj() to allow resctrl to walk the kobjects in the
kset.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>