untested: iommu/arm-smmu-v3: Register SMMU capabilities with MPAM
commit 2d52f0fef96b24fc0ba4e05a1e966c069f50a83e from
git@git.assembla.com:cavium/WindRiver.linux.git

Traffic in the system can be tagged with a PARTID and PMG. Different
requestors can support a different number of bits for these fields.

Before MPAM can be used, the MPAM driver has to discover the minimum
number of bits supported by any requestor, which affects the range
of PARTID and PMG that can be used.

Detect whether the SMMU supports MPAM, if it does provide the MPAM
driver with the maximum PARTID and PMG values.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>