LF-4836 arm64: imx8mx vpu: make sure output streamon cmd is seet to daemon after buffer is queued
commit  be773ff689345bf2d007b8eb58bbac5f8bc3fcf7 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

the cmd V4L2_DAEMON_VIDIOC_STREAMON_OUTPUT should be sent to daemon
when the minimum needed buffers are queued.
so we should make sure that when buffers are queued, we send it to daemon.
These is a bug that driver may not send it to daemon due to some statue
issue.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>