MLK-20697-1: drm/panel: panel-raydium-rm67191: Add MIPI_DSI_MODE_NO_EOT_PACKET flag
commit  0aedb2f15d419d38e9e16c0f2bb0152a48657bba from
https://github.com/nxp-imx/linux-imx lf-6.12.y

This flag was automatically used in the DSI driver, making this panel
work, but the way that flag was used was wrong. Since, fixing the use of
this flag in the DSI driver now breaks this panel, just add it here.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>