arm_mpam: resctrl: Call resctrl_exit() in the event of errors
commit 989e9395917e491702a8741b8c8e080942df3840 from
git@git.assembla.com:cavium/WindRiver.linux.git

All of MPAMs errors indicate a software bug, e.g. an out-of-bounds
partid has been generated. When this happens, the mpam driver
is disabled.

If resctrl_init() succeeded, also call resctrl_exit() to
remove resctrl.

If the filesystem was mounted in its traditional place, it is
no longer possible for processes to find it as the mount point
has been removed. If the filesystem was mounted elsewhere, it
will appear that all CPU and domains are offline. User-space
will not be able to update the hardware.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>