LF-4972: ASoC: fsl_rpmsg: allocate a smaller buffer size for capture
commit  4336b828c7ce15a7e4df7ac15cf61749e8189996 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Current the capture size is 0x6000000, plus playback size (0x6000000)
the total size is exceed the maximum buffer size for the sound card.

There are several buffer size setup modes for different rpmsg
soundcards. For rpmsg-micfil soundcard, the capture buffer size is
0x100000. For rpmsg-audio soundcard, the normal playback buffer size
as well as capture is 0x10000. In the case that lpa is enabled in
rpmsg-audio soundcard, the playback buffer size is set to 0x6000000
and the capture buffer size is 0x100000.

Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Chancel Liu <chancel.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>