LF-5173-4 drm/imx: lcdif: crtc: Make it possible to disable entire display pipeline in self refresh mode
commit  6ed112e298e8106403cfb7230aa139bba6e1e143 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

In the self refresh mode, the DRM atomic helper drm_atomic_crtc_needs_modeset()
would refuse to disable relevant encoder and bridges if the active flag of
the next new CRTC state is set to false, which implies the entire display
pipeline is to be disabled, e.g., the fb emulation is to be blanked.
Forcing the connectors_changed flag of the CRTC state to true in the
->atomic_check() callback in this case makes it possible to disable the
encoder and bridges as the DRM atomic helper references that flag to
determine if the encoder and bridges need to be disabled.

This patch fixes the below patch:
LF-4887-6 drm/imx: lcdif: Add MPU interface(8080 mode) support

Cc: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>