drm/bridge: adv7511: fix of_reconfig
commit  6e5d40b690aa3fd0f8ffc994d5fd8174753874b6 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

If probe fails, this driver will remove its dts connection with the
dsi host, so that the host won't try to add it to the component
framework (if used). But, if the return code is -EPROBE_DEFER, skip this
logic so the adv driver can try probing again and retry the dsi host
connection next time.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>