octeontx2-pf: Limit number of QOS queues
commit 36264a0ed3d5491744ed4013543989f24e74a5d2 from
git@git.assembla.com:cavium/WindRiver.linux.git

Receive/trasmit irqs and cpus are one to one mapped for better
performace. When number of cpus are lesser than Qos queues the
msi vector allocation is failing to fix the issue  number of
Qos queues are limited to number of cpus.

Change-Id: I4af42fb26e67a370f75dc688f15bc6db39a00595
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/72643
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>