dpaa2-eth: increase busy retries when interracting with QBMAN
commit  13eeef40295e4901c95bc8025e3dd5bf43885204 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

It seems that there are circumstances when access to QBMAN through the
software portals will be delayed. Accessing some lower speeds interfaces
while also QBMAN commands are issued from the kernel will lead to
software timeouts happening in the dpaa2-eth driver.

What we have observed is that management commands like re-arming the
interrupts on a specific channel, waiting for a dequeue response to be
available etc, will take a longer time to complete.
All these commands have to wait for a valid bit to be set for the
command to be interpreted as successfully completed.

Increase the maximum number of times the Linux kernel drivers will busy
poll for a successful result of one of these commands.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>