LF-6626 drm/bridge: synopsys: dw-mipi-dsi: Remove bridge in __dw_mipi_dsi_remove()
commit  a9d11f8d77f6ef55ccd6611ff3da0b01cb1676d9 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The bridge is added in __dw_mipi_dsi_probe(), so it should be
removed in  __dw_mipi_dsi_remove().  Otherwise, it will be left
in the global bridge list and leaked to drivers which try to
find the bridge.

Fixes: c206c7faeb32 ("drm/bridge: dw-mipi-dsi: Find the possible DSI devices")
Cc: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>