LF-7107-2: media: capture: ov5640: Reduce the delay time for sensor to be stable
commit  0d2303c282c2dae8ded90bfb89230cf9170e6c21 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

When do stress test of capture, taking pictures and rebooting.
it will find that about 5% of them are black, the pictures shouldn't be black.

Root cause:
The delay time in the initial of the ov5640 wouldn't prevent dark image from
passing to the buffer. So it need to be optimized.

The few frames of ov5640 AE/AG is not stable,
and at the begin of the few frames, the ov5640 sensor 3A is under adjustment.
And the picture data is dirty, should be skipped.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>