LF-7157 drm/imx: dpu: crtc: Disable plane(s) on-the-fly properly in ->atomic_flush()
commit  844d4b317d28041800f3630ffd55df8e42cb024f from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Similar to ->atomic_disable(), it appears that we have to ensure not to
relinquish CPU by disabling local irq and preemption during the procedure
of waiting for FrameGen frame counter moving, DPRC repeat_en disablement
and ExtDst shadow load initiation so that plane(s) backed by prefetch
engines and fetchunits can be disabled on-the-fly properly.  Otherwise,
there is chance that the plane(s) are not correctly disabled with the
warnings like below on some SoCs:

[   27.770949] imx-dpu-crtc imx-dpu-crtc.0: flush - wait for content shdld done timeout
[   27.778764] imx-dpu-crtc imx-dpu-crtc.0: flush - FrameGen requests to read empty FIFO

While at it, the FIXME tag in ->atomic_flush() is dropped.

Reported-by: Justin Jiang <justin.jiang@nxp.com>
Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Tested-by: Justin Jiang <justin.jiang@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>