LF-6930 [#imx-2969] Skip checking MC idle bits in isGPUIdle function on 8ulp
commit  9511a3c31ec9fd34e82e191f9929dd33c6b543d5 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Software workaround for MC idle bits not set for a long time on 8ULP.
The reason is MC idle status depend on the undone memory request counter,
which cannot decrease to zero due to HW bug of 8ULP GPU. Actually all
modules finished the job and become idle properly. MC not-idle is a
false alert if there is no HW activity in the whole pipeline. So that
it is safe to skip the check for MC idle bit in SW workaround.

Signed-off-by: Xiaojun.Chen <xiaojun.chen@versilicon.com>
Reviewed-by: Guangliu Ding <guangliu.ding@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>