arm64: dts: ls1028a-rdb-dpdk: remove pause property
commit  dabf93ff68d122192132172c0664fa01b0dbea46 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

DPDK is not capable to handle any type of congestion
or back-pressure. So if there is congestion on the network,
HW can generate PAUSE frames to limit the traffic and packets
can be dropped in the DPDK.

So in this patch, removing the pause property from switch and
ethernet ports.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>