drivers: i2c: octeon: Avoid potential infinite loop
commit 624fbf287969c90afa589b9a0449d10c93bad41a from
git@git.assembla.com:cavium/WindRiver.linux.git

The time_left variable is used and set to 1000, which is not
being checked in while loop condition - this could trigger as
infinite loop in case the test flag is not set. Adding time_left
check in while condition as a way to exit the loop in that case.

Signed-off-by: Piyush Malgujar <pmalgujar@marvell.com>
Change-Id: I0a383045fc5144da9a64385602724fd88facf8ce
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>