LF-7857-2: arm64: dts: imx8mn-evk: Modify the PHY node for new BB
commit  2ed85eb8acc8c36790d7ba9d3520982530b18b2b from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The base board of i.MX8MM/MN EVK has been re-designed due to EOL
component for the ethernet PHY. And the new PHY is RTL8211. In
order to support both the new and old base board with the same
software, we add the properties of RTL8211 to the phy node. But
notice that, this is a workaround and not "upstreamable". We will
find a better method to the upstream in future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>